diff, pragmas and unsigned long -> size_t
* Functions that returned their size by reference (map, set and graph)
as unsigned long now return as size_t, which is semantically more fitting
and may occupy less space (in 32-bit systems, it would occupy 4 bytes, instead
of 8 bytes, which is the size of a long).
* Added token 'z' to testerlib that parses size_t arguments
* Rewritten some tester code to parse size_t, instead of long arguments
* pygraph now returns its size by a separate method, since seq_length returns
Py_ssize_t, which is an opaque type for ssize_t (signed size_t), which can have
overflow issues with size_t. Therefor, another function was created, returning
a PyObject wrapping a size_t value.
* Added pragmas to ignore some MSVC warnings
* Changed diff algorithm and return value

+++ graph/graph.h
@@ -109,7 +109,7 @@ GraphReturnID graphIsDirected(Graph *pGraph, int *pIsDirected);
 * 	- "pGraph" is NULL
 * 	- "pSize" is NULL
 */
-GraphReturnID graphGetNumberOfVertices(Graph *pGraph, unsigned long *pSize);
+GraphReturnID graphGetNumberOfVertices(Graph *pGraph, size_t *pSize);
 
 /**
 * Get next vertex in graph (loops)
@@ -136,7 +136,7 @@ GraphReturnID graphGetNextVertex(Graph *pGraph, void **pV);
 * 	- vertex "v" does not exist
 */
 GraphReturnID graphGetVertexOutDegree(Graph *pGraph, void *v,
-	unsigned long *pOut);
+	size_t *pOut);
 
 /**
 * Get in degree of a given vertex, that is, how many edges go to it
@@ -151,7 +151,7 @@ GraphReturnID graphGetVertexOutDegree(Graph *pGraph, void *v,
 * 	- vertex "v" does not exist
 */
 GraphReturnID graphGetVertexInDegree(Graph *pGraph, void *v,
-	unsigned long *pIn);
+	size_t *pIn);
 
 /**
 * Get degree of a given vertex, that is, how many edges are incident to it
@@ -168,7 +168,7 @@ GraphReturnID graphGetVertexInDegree(Graph *pGraph, void *v,
 * 	- vertex "v" does not exist
 */
 GraphReturnID graphGetVertexDegree(Graph *pGraph, void *v,
-	unsigned long *pDegree);
+	size_t *pDegree);
 
 /**
 * Get next neighbour of given vertex
